bug #436726 reported by paddy_mul: Ajax support for client changing.
rsevero had done a 6.x patch, but I did not base my changes off of it.

I tried to do my best at handling client changes gracefully, but just in case, I added some validation.

I cleaned up the way the stuff attaches to the form, but it is still doing a lot of weird stuff.